% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_character_plate_pos.R
\name{as_character_plate_pos}
\alias{as_character_plate_pos}
\title{Converteer numerieke positie op een plaat een letter-cijfer codering}
\usage{
as_character_plate_pos(pos, lanes = 8)
}
\arguments{
\item{pos}{positie op de plaat}

\item{lanes}{aantal kolommen op de plaat (op 96-well plaat is dit meestel 8)}
}
\value{
character vector met posities in het formaat A1, B2, C12
}
\description{
Converteer numerieke positie op een plaat een letter-cijfer codering
}
\examples{
as_character_plate_pos(c(1,37,38,96))
}
