% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii_convert.R
\name{asc_to_int}
\alias{asc_to_int}
\title{Convert ascii characters to their ascii number}
\usage{
asc_to_int(x, simplify = FALSE)
}
\arguments{
\item{x}{vector of strings}

\item{simplify}{argument as used in the sapply function}
}
\value{
integer numbers of the strings. One entry per letter, one column per string
}
\description{
Convert ascii characters to their ascii number
}
