% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_parse_xls_default.R
\name{batch_parse_xls_default}
\alias{batch_parse_xls_default}
\title{Parse default excel bestanden}
\usage{
batch_parse_xls_default(file, sheet = "LIMSIMPORT", cell = "B1", user = "")
}
\arguments{
\item{file}{file die geopend moet worden}

\item{sheet}{het tabblad dat ingelezen moet worden}

\item{cell}{de eerste cel die wordt ingelezen}

\item{user}{gebruikersnaam die de opdracht heeft gegeven}
}
\value{
dataset
}
\description{
Parse default excel bestanden
}
