% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_parse_xls_tripledash.R
\name{batch_parse_xls_tripledash}
\alias{batch_parse_xls_tripledash}
\title{Lees file in waarbij stalen met: text_id---testnummer worden gebruikt}
\usage{
batch_parse_xls_tripledash(
  file,
  sheet = "LIMSIMPORT",
  cell = "B1",
  maxcol = 0,
  user = "TEST"
)
}
\arguments{
\item{file}{file die ingelezen moet worden}

\item{sheet}{naam van de sheet (case sensitive)}

\item{cell}{eerste cel}

\item{maxcol}{maximaal aantal kolommen}

\item{user}{LIMS gebruikers naam van de inlezer}
}
\value{
dataset
}
\description{
Lees file in waarbij stalen met: text_id---testnummer worden gebruikt
}
