% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elc_functions_obsolete.R
\name{calc_elc_stats}
\alias{calc_elc_stats}
\title{Bereken de ELC basisstatistieken van een numerieke kolom (OUDE MANIER)}
\usage{
calc_elc_stats(x)
}
\arguments{
\item{x}{numerieke vector met waarden}
}
\value{
data.frame met basisstatistieken
}
\description{
Bereken de ELC basisstatistieken van een numerieke kolom (OUDE MANIER)
}
