% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elc_calculate_elc_periodic_stats.R
\name{calculate_elc_periodic_stats}
\alias{calculate_elc_periodic_stats}
\title{Calculate stats for ELC Yearchart}
\usage{
calculate_elc_periodic_stats(data)
}
\arguments{
\item{data}{dataset in the specific format used in elc periodic}
}
\value{
dataset with base stats like n, mean sd, the same qfter removing outliers en the vqlues the previous period
}
\description{
Calculate stats for ELC Yearchart
}
