% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber_test_functions.R
\name{check_db}
\alias{check_db}
\title{Check if db connection is alive}
\usage{
check_db(env = "PRD")
}
\arguments{
\item{env}{PRD, UAT of DEV}
}
\value{
list with result of check
}
\description{
Check if db connection is alive
}
