% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber_test_functions.R
\name{check_multiply}
\alias{check_multiply}
\title{Simple multiplication to test service}
\usage{
check_multiply(a, b)
}
\arguments{
\item{a}{numeric-convertible}

\item{b}{numeric-convertible}
}
\value{
numeric value
}
\description{
Simple multiplication to test service
}
