% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_ascii_convert.R
\name{convert_to_simple_ascii}
\alias{convert_to_simple_ascii}
\title{Remove special characters from string}
\usage{
convert_to_simple_ascii(
  x,
  keep_codes = c(38:128, 181),
  replace = data.frame(181, 181)
)
}
\arguments{
\item{x}{vector of strings containing special characters}

\item{keep_codes}{ascii-codes to be kept. By default 38:128 and 181}

\item{replace}{2-kolom data.frame, in which the character in the first column is replaced by the character in the second column}
}
\value{
ascii vector without special characters
}
\description{
Remove special characters from string
}
