% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzy_copy_and_update_scripts.R
\name{copy_and_update_scripts}
\alias{copy_and_update_scripts}
\title{Copy package scripts and resources to a single LIMS path if they are outdated.}
\usage{
copy_and_update_scripts(target_path)
}
\arguments{
\item{target_path}{The single LIMS directory path to check and update.}
}
\value{
A single logical value: `TRUE` if scripts were successfully updated
  or were already up-to-date, `FALSE` on failure.
}
\description{
This function compares MD5 hashes of scripts in the package's 'inst/lims_scripts'
directory with those in the target LIMS path. If they differ, it backs up the
existing scripts and copies the new ones over. It also copies the 'resources'
folder and its contents to the target path.
}
