% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_scripts_to_lims_path.R
\name{copy_scripts_to_lims_path}
\alias{copy_scripts_to_lims_path}
\title{Kopieer R scripts naar LIMS directory}
\usage{
copy_scripts_to_lims_path(path = "L:\\\\LabWare-7\\\\Data\\\\R_SCRIPTS")
}
\arguments{
\item{path}{de folder waar de R scripts terecht moeten komen}
}
\value{
vector met status per bestand of het succesvol gekopieerd is. Zou allemaal TRUE moeten zijn
}
\description{
De bestanden worden vanuit het inbolimsintern pakket gekopieerd naar het gekozen pad. De bestaande bestanden in dit pad, worden verplaatst naar een backup folder (met datumaanduiding) in het pad
}
