% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_update_functions.R
\name{db_update_get_data}
\alias{db_update_get_data}
\title{Interpreteer data en zet om naar vorm van C_REQUESTED_DB_UPDATES}
\usage{
db_update_get_data(data, conn)
}
\arguments{
\item{data}{data ingelezen uit een csv die omgezet moet worden (1 rij per dataset want gebruikt via rowwise)}

\item{conn}{DBI connectie naar LIMS databank}
}
\value{
data.frame omzetbaar naar queries voor C_REQUESTED_DB_UPDATES
}
\description{
Interpreteer data en zet om naar vorm van C_REQUESTED_DB_UPDATES
}
