% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_update_functions.R
\name{db_update_worklist_table}
\alias{db_update_worklist_table}
\title{Maak update queries zodat LIMS die kan verwerken}
\usage{
db_update_worklist_table(conn, data)
}
\arguments{
\item{conn}{dbi connectie naar db}

\item{data}{data ingelezen uit file die in bruikbaar formaat omgezet moet worden}
}
\value{
list met update queries
}
\description{
Maak update queries zodat LIMS die kan verwerken
}
