% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elc_functions.R
\name{elc_fixlim_data}
\alias{elc_fixlim_data}
\title{Genereer ELC data om te archiveren als plot in de lims db}
\usage{
elc_fixlim_data(
  plotdata,
  colors = c("lightblue3", "green4", "orange", "red", "blue4"),
  base_size = 1.5,
  digits = 5
)
}
\arguments{
\item{plotdata}{te plotten data die BATCHNR, ORDER, ENTRY bevat}

\item{colors}{basiskleuren voor de kaart}

\item{base_size}{basisgroote van de bollen}

\item{digits}{aantal cijfers na de komma behouden}
}
\value{

}
\description{
Genereer ELC data om te archiveren als plot in de lims db
}
