% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elc_elc_htest.R
\name{elc_htest}
\alias{elc_htest}
\title{Elementaire testen (t-test en variantie test)}
\usage{
elc_htest(data, valuecol = "WAARDE", grpcol = "grp", label = "")
}
\arguments{
\item{data}{data waarin de data voor de t-test en F-test te vinden zijn}

\item{valuecol}{kolomnaam voor de waarden}

\item{grpcol}{kolomnaam voor de groepen}

\item{label}{identificatielabel voor resultaten}
}
\value{
data.frame met resultaten t-test en var.test
}
\description{
Elementaire testen (t-test en variantie test)
}
