% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elc_shewhart_plot.R
\name{ELC_shewhart_plot}
\alias{ELC_shewhart_plot}
\title{Plot ELC shewhart plot}
\usage{
ELC_shewhart_plot(
  subdata,
  borders = NULL,
  base_color = "lightblue3",
  max_s_plot = 5,
  interactive = FALSE,
  title = NULL,
  fig_height = 600
)
}
\arguments{
\item{subdata}{data om te plotten met kolommen batchnr, waarde, color, size}

\item{borders}{dataframe met de kolommen val en color om horizontale lijnen voor de s-grenzen te trekken}

\item{base_color}{basiskleur in de grafieken}

\item{max_s_plot}{Op hoeveel s moet de plot alles erbuiten niet meer tonen. Indien 0 of NA toon de hele plot}

\item{interactive}{Moet de plot een interactief plotly object worden?}

\item{title}{De titel die aan de plot toegevoegd wordt}

\item{fig_height}{De hoogte van de figuren in pixels}
}
\value{
ggplot2 object
}
\description{
Plot ELC shewhart plot
}
\examples{
{
g <- "green4"; b <- "lightblue3"; r = "red"
subdata <- data.frame(batchnr = c(1,1,1,2,3,4,5,5,5,6,7,8,8,9,10),
waarde = c(10,-8,20,32,-16,-36,63,16,8,21,1,8,14,-3,9),
color = c(g,b,b,r,g,r,r,b,b,g,g,g,b,g,g),
size = c(1,NA,NA,2,1,2,NA,NA,NA,1,1,1,NA,1,1),
UNITS = "m",
BATCH = LETTERS[1:15],
ORDER = 1:15,
eval = c(1,NA,NA,1,1,1,1,NA,NA,1,1,1,NA,1,1))
borders <- data.frame(lim = -3:3, val = (-3:3)*10,
color = c("red", "gold", "green4", "lightblue3", "green4", "gold", "red"))
ELC_shewhart_plot(subdata, borders, max_s_plot = 4)
ELC_shewhart_plot(subdata, borders, max_s_plot = 0)
ELC_shewhart_plot(subdata, borders, max_s_plot = 4, interactive = TRUE)
}

}
