% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_ms_create_plates.R
\name{gen_ms_create_plates}
\alias{gen_ms_create_plates}
\title{Randomiseer data, kies substalen, en maak de platen aan}
\usage{
gen_ms_create_plates(DNA, ...)
}
\arguments{
\item{DNA}{dataset met DNA_RUN_ID (1 cijfer die voor dit project gelijk blijft), SAMPLE_NUMBER (de sample_numbers uit LIMS), SAMPLE_TYPE (N = normaal sample, Q = QCmethod)}

\item{...}{argumenten die doorgegeven worden aan de functie \link{randomiseCapilar}}
}
\value{
dataset met naast de velden in DNA, de extra velden PLATE_SEQ (nummer van de plaat), PLATE_POSITION (positie op de plaat) en REP_SAMPLE_NUMBER (staalnummer waarvan dit een substaal is). Deze dataset moet per definitie meer rijen bevatten dan DNA, aangezien er replicates gelogd worden, en nieuwe controlestalen en blanco's indien de stalen over meerdere platen worden verdeeld
}
\description{
Randomiseer data, kies substalen, en maak de platen aan
}
