% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_seq_create_plates.R
\name{gen_seq_create_plates}
\alias{gen_seq_create_plates}
\title{Plaatdata aanmaken voor sequentie-analyse.}
\usage{
gen_seq_create_plates(
  DNA,
  params,
  qcmpos = 37,
  blancpos = NULL,
  platesize = 96
)
}
\arguments{
\item{DNA}{dataset met DNA_RUN_ID (1 cijfer die voor dit project gelijk blijft), SAMPLE_NUMBER (de sample_numbers uit LIMS), SAMPLE_TYPE (N = normaal sample, Q = QCmethod)}

\item{params}{de dataset met de parameters die vanuit het lims werden doorgegeven}

\item{qcmpos}{positie voor de qc method}

\item{blancpos}{Positie voor de blanco. NULL indien geen blanco's nodig zijn}

\item{platesize}{de grootte van de plaat}
}
\value{
dataset met naast de velden in DNA, de extra velden PLATE_SEQ (nummer van de plaat), PLATE_POSITION (positie op de plaat) en REP_SAMPLE_NUMBER (staalnummer waarvan dit een substaal is). Deze dataset moet per definitie meer rijen bevatten dan DNA, aangezien er replicates gelogd worden, en nieuwe controlestalen en blanco's indien de stalen over meerdere platen worden verdeeld
}
\description{
Plaatdata aanmaken voor sequentie-analyse.
}
