% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_chart_functions.R
\name{generate_html_source}
\alias{generate_html_source}
\title{Genereer basis html file}
\usage{
generate_html_source(path, entries, show_pagesource = FALSE, ...)
}
\arguments{
\item{path}{de volledige padnaam naar de html file}

\item{entries}{list met textstrings die de body van de html uitmaken}

\item{show_pagesource}{logische waarde om te bepalen of de html source op het scherm getoond moet worden}

\item{...}{andere parameters (nog niet in gebruik)}
}
\value{
html source
}
\description{
Genereer basis html file
}
