% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filegrabber_batch_functions.R
\name{get_batch_info}
\alias{get_batch_info}
\title{Verkrijg de batchinformatie uit de LIMS DB}
\usage{
get_batch_info(conn, batchname)
}
\arguments{
\item{conn}{dbi connectie naar de lims databank}

\item{batchname}{string met de naam van de batch}
}
\value{
dataset met 1 rij en volgende kolommen: name, template,  template_version, owner, c_import_routine, c_import_sheet
}
\description{
Verkrijg de batchinformatie uit de LIMS DB
}
