% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filegrabber_batch_functions.R
\name{get_batchname_from_file}
\alias{get_batchname_from_file}
\title{Haal de batchnaam uit een importfile}
\usage{
get_batchname_from_file(file)
}
\arguments{
\item{file}{de naam van een bestand met importresultaten, de batch wordt gevonden door te zoeken naar ---, alles daarvoor wordt als batchnaam beschouwd}
}
\value{
character string met de batch naam
}
\description{
Haal de batchnaam uit een importfile
}
