% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_env.R
\name{get_current_environment}
\alias{get_current_environment}
\title{Detect the current operating environment (DEV, UAT, or PRD or LW7)}
\usage{
get_current_environment(odbc)
}
\arguments{
\item{odbc}{the name of the odbc passed on by commandArgs(trailingOnly = TRUE)[1]}
}
\value{
A string: "DEV", "UAT", or "PRD" or "LW7". Defaults to "PRD" with a warning
  if no specific environment can be identified.
}
\description{
This function attempts to identify the environment by checking system
variables, hostnames, or known directory structures.
}
