% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_get_edna_plate_data.R
\name{get_edna_plate_data}
\alias{get_edna_plate_data}
\title{Zet projectstalen klaar voor op de plaat}
\usage{
get_edna_plate_data(data, rep_fractie = 0.1, plate_size = 96,
  column_size = 8, n_qcm = 1, n_blank = 3, max_reps = 2,
  min_filled = 24)
}
\arguments{
\item{data}{datset met projectstalen. Bevat minstens de kolom DNA_ID, DNA_RUN_ID, SAMPLE_NUMBER, REP_SAMPLE_NUMBER}

\item{rep_fractie}{welk aandeel van de stalen (tussen 0 en 1) wordt gebruikt om substalen te maken}

\item{plate_size}{grootte van de plaat. Typisch 96}

\item{column_size}{lengte van de plaatkolommen. Typisch 8 voor een 96-well plaat}

\item{n_qcm}{aantal qc methods per plaat}

\item{n_blank}{aantal blanco's per plaat}

\item{max_reps}{maximaal aantal keer dat een staal gerepliceerd mag worden in de plaat. Dus elk van de rep_fractie stalen kan maximaal max_reps keer als substaal voorkomen. In het totaal komt het staal dus maximaal  1 + max_reps keer voor op de plaat}

\item{min_filled}{hoeveel elementen van een plaat moeten gevuld worden, om een nieuwe plaat waard te zijn. Indien de overflow van de vorige plaat minder dan 24 is, wordt het aantal substalen gereduceerd, zodat geen nieuwe plaat nodig is.}
}
\value{
dataset die klaargezet is om in de functie set_edna_plate_positions te gebruiken
}
\description{
Zet projectstalen klaar voor op de plaat
}
