% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elc_get_ELC_data.R
\name{get_ELC_data}
\alias{get_ELC_data}
\title{Verkrijg de ELC data}
\usage{
get_ELC_data(dbcon, sqlfile, keep = 30, productversions = NULL, logfile = NULL)
}
\arguments{
\item{dbcon}{dbconnection object (DBI)}

\item{sqlfile}{path naar de file die de sql code bevat}

\item{keep}{aantal batches te behouden voor de figuur}

\item{productversions}{data.frame me een kolom PRODUCT en VERSIE om te bepalen welke limieten gebruikt moeten worden. IF NULL dan worden de maximale versies gebruikt voor elk product}

\item{logfile}{indien niet NULL schrijf debugging info weg in logfile}
}
\value{
dataset met alle te verwerken gegevens
}
\description{
Verkrijg de ELC data
}
