% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_css_functions.R
\name{get_embedded_image}
\alias{get_embedded_image}
\title{Encode a package image to Base64}
\usage{
get_embedded_image(filename)
}
\arguments{
\item{filename}{Name of the image file in inst/images/}
}
\description{
Encode a package image to Base64
}
