% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_importfile_parse.R
\name{get_file_extension}
\alias{get_file_extension}
\title{Verkrijg de bestandsextensie}
\usage{
get_file_extension(file)
}
\arguments{
\item{file}{de naam van de file waarvoor de extensie gezocht wordt. De extensie is alles wat na het laatste '.' komt}
}
\value{
character string met de extensie
}
\description{
Verkrijg de bestandsextensie
}
