% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_print_labels.R
\name{get_label_format}
\alias{get_label_format}
\title{Get Label Format Lines from LIMS}
\usage{
get_label_format(conn, format_name)
}
\arguments{
\item{conn}{A DBI connection object (e.g., from `lims_connect()`).}

\item{format_name}{Character. The unique name of the format in `T_LABEL_FORMAT`.}
}
\value{
A dataframe containing the layout instructions for the label.
}
\description{
Retrieves the specific design lines (fields, coordinates, fonts) for a given label format.
}
\examples{
\dontrun{
conn <- limsdb_connect()
format_lines <- get_label_format(conn, "PROJ_LABEL_EXTID")
}
}
