% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get_lims_constant.R
\name{get_lims_constant}
\alias{get_lims_constant}
\title{Get a value from system configuration tables}
\usage{
get_lims_constant(conn, name)
}
\arguments{
\item{conn}{odbc connection object}

\item{name}{The name of the constant or entry_code to retrieve.}
}
\value{
Character value of the first match found, or NULL.
}
\description{
Searches for a named value first in LIMS_CONSTANTS and then in CONFIG_SYS.
}
\examples{
\dontrun{
# This will search for 'DEFAULTDIR_LAB_GENERIC' in both tables
get_system_value(conn, "DEFAULTDIR_LAB_GENERIC")
}
}
