% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_chart_functions.R
\name{get_run_chart_data}
\alias{get_run_chart_data}
\title{Get the data voor runchart}
\usage{
get_run_chart_data(conn, qry)
}
\arguments{
\item{conn}{db connectie}

\item{qry}{uit te voeren query}
}
\value{
dataset klaar om een run chart van te plotten
}
\description{
Get the data voor runchart
}
