% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_update_functions.R
\name{get_sample_and_test_number}
\alias{get_sample_and_test_number}
\title{Verkrijg sample_number en test_number op basis van text_id en analysenaam}
\usage{
get_sample_and_test_number(conn, text_id, extern_id = NULL, analysis = NULL)
}
\arguments{
\item{conn}{DBI connectie naar LIMS DB}

\item{text_id}{text_id waarvoor sample nummer gezocht moet worden}

\item{extern_id}{optioneel extern_id waar ook op gecheckt moet worden voor extra veiligheid}

\item{analysis}{optioneel enkel wanneer een specifieke test opgehaald moet worden}
}
\value{
data.frame met sample_number, sample_id en test_number als velden
}
\description{
Verkrijg sample_number en test_number op basis van text_id en analysenaam
}
