% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcc_gg_lims_shewhart.R
\name{gg_lims_shewhart}
\alias{gg_lims_shewhart}
\title{Shewhart plot}
\usage{
gg_lims_shewhart(
  plotdata,
  red_rules = 1:2,
  orange_rules = 3:6,
  yellow_rules = 7:8,
  ...
)
}
\arguments{
\item{plotdata}{a dataset typically obtained through the function lims_shewhart}

\item{red_rules}{numeric vector of rule violations that should be marked as red}

\item{orange_rules}{numeric vector of rule violations that should be marked as orange}

\item{yellow_rules}{numeric vector of rule violations that should be marked as yellow}

\item{...}{not currently in use}
}
\value{
Shewhart ggplot
}
\description{
Shewhart plot
}
