% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcc_html_qc_report.R
\name{html_qc_report}
\alias{html_qc_report}
\title{Maak een HTML rapport}
\usage{
html_qc_report(data, path = "", digits = c(5, 5))
}
\arguments{
\item{data}{dataset verkregen na de functie select_control_samples}

\item{path}{pad indien afwijkend van de werkdirectory. Zeker eindigen met "/". Zorg dat deze directory reeds bestaat.}

\item{digits}{vector met 2 elementen, het eerste hoeveel digits in de data, het tweede hoeveel in de samenvattende statistieken}
}
\value{
HTML rapport met QC kaarten
}
\description{
Maak een HTML rapport
}
\examples{
\dontrun{
batch = "IC_AN-190430-1"
analysis = "IC_ANIONEN"
components = c("NO2", "NO3")
testdf <- select_control_samples(conn, num = 30, batch, analysis, components)
html_qc_report(testdf)
}
}
