% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_chart_functions.R
\name{html_run_chart}
\alias{html_run_chart}
\title{Maak een html met de run chart}
\usage{
html_run_chart(plotdata, project, path, split_fig = TRUE, ...)
}
\arguments{
\item{plotdata}{data om te plotten}

\item{project}{projectnaam}

\item{path}{basispad waarin geschreven wordt}

\item{split_fig}{logische waarde die bepaalt of de plots ook uitgesplitst per variabele getoond worden}

\item{...}{andere argumenten voor subfuncties zoals show_pagesource voor generate_html_source}
}
\value{
naam van de gerenderde html file
}
\description{
Maak een html met de run chart
}
