% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_chart_functions.R
\name{html_run_chart_embedded}
\alias{html_run_chart_embedded}
\title{Generate and Embed Run Charts into an HTML String}
\usage{
html_run_chart_embedded(
  plotdata,
  project,
  split_fig = TRUE,
  chart_header = "",
  ...
)
}
\arguments{
\item{plotdata}{A data frame with the necessary columns for plotting.}

\item{project}{A character string for the plot title.}

\item{chart_header}{A character string for the HTML header.}

\item{...}{Not used.}
}
\value{
A character string containing a self-contained HTML snippet with embedded charts.
}
\description{
Generate and Embed Run Charts into an HTML String
}
