% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_ascii_convert.R
\name{int_to_asc}
\alias{int_to_asc}
\title{Convert integers to their ascii symbols}
\usage{
int_to_asc(x)
}
\arguments{
\item{x}{integer value to be converted to the corresponding ascii symbol}
}
\value{
list with ascii numbers for each character
}
\description{
Convert integers to their ascii symbols
}
