% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elc_qcc_lims_shewhart.R
\name{lims_shewhart}
\alias{lims_shewhart}
\alias{lims_shewhart_default}
\alias{lims_shewhart.lims_data}
\alias{lims_shewhart.default}
\title{Calculate Shewhart rules (generic)}
\usage{
lims_shewhart(x, ...)

\method{lims_shewhart}{lims_data}(x, entrycol = "ENTRY", ...)

\method{lims_shewhart}{default}(x, center = NULL, sd = NULL, rules = lims_shewhart_rules, ...)
}
\arguments{
\item{x}{lims_data object containing all info to make a shewhart plot}

\item{...}{other variables passed on}

\item{entrycol}{the column name in the dataset that is used as value column for the results}

\item{center}{fixed mean value, or calculated when set to NULL}

\item{sd}{fixed standard deviation, or calculated when set to NULL}

\item{rules}{definition of rules ??? ADD MORE EXPLANATION ???}
}
\value{
dataset with the violations for all the 8 shewhart rules
}
\description{
Calculate Shewhart rules (generic)

Calculate Shewhart rules (lims_data object)

Calculate Shewhart rules (default)
}
