% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcc_lims_shewhart_rules.R
\name{lims_shewhart_rules}
\alias{lims_shewhart_rules}
\title{Wrapper function that calculates all 8 Shewhart rules for lab results}
\usage{
lims_shewhart_rules(x, center, sd, runs = c(1, 9, 6, 14, 3, 5, 15, 8),
  ...)
}
\arguments{
\item{x}{numeric vector of lab results}

\item{center}{mean value around which all rules are checked}

\item{sd}{standard deviation used as basis for the rule violations}

\item{runs}{vector of 8 elements that contain the run before violation is triggerd for all 8 rules. The first should always be exactly 1, the rest larger than 1.}

\item{...}{not used}
}
\value{
matrix containing the value of the lab result, with all 6-sigma stats (-3s, -s, -1s, center, +1s, +2s, +3s) and all 8 rule violations as a TRUE/FALSE vector
}
\description{
Wrapper function that calculates all 8 Shewhart rules for lab results
}
