% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elc_functions.R
\name{list_all_qc_data}
\alias{list_all_qc_data}
\title{Lijst alle controlestalen waarvoor limieten bestaan op binnen de gekozen periode}
\usage{
list_all_qc_data(conn, min_date = "2020-01-01", max_date = "2021-01-01")
}
\arguments{
\item{conn}{connectie naar db object}

\item{min_date}{vroegste beschouwde datum yyyy-mm-dd}

\item{max_date}{laatst beschouwde datum yyyy-mm-dd}
}
\value{
dataset met alle resultaatinformatie voor de geselecteerde controlestalen
}
\description{
Lijst alle controlestalen waarvoor limieten bestaan op binnen de gekozen periode
}
