% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_env.R
\name{load_encrypted_renviron}
\alias{load_encrypted_renviron}
\title{Load environment variables from an encrypted .Renviron file}
\usage{
load_encrypted_renviron(path)
}
\arguments{
\item{path}{optional path (if not provided, R will search for it)}
}
\value{
Invisibly returns `TRUE` on success or `FALSE` on failure.
}
\description{
This function searches for a '.Renviron.enc' file in the current working
directory and the user's home directory, decodes its Base64 content,
and loads the variables for the current R session.
}
