% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfile_functions.R
\name{logfile_start}
\alias{logfile_start}
<<<<<<< HEAD
\title{Make name of log file based on current datetime, directory, prefix and suffix}
=======
\title{Creert logbestand (opent connectie)}
>>>>>>> d798110a5b4d52f71d723703657283586349b6aa
\usage{
logfile_start(path = "<auto>", prefix = "", suffix = "")
}
\arguments{
\item{path}{directory waar de logfile wordt aangemaakt . <auto> gebruikt de default}

\item{prefix}{vorvoegsel in de file, bijvoorbeeld welk script gebruikt werd}

<<<<<<< HEAD
\item{suffix}{suffix tot the file to make it recognisable}
}
\value{
file connection
}
\description{
Make name of log file based on current datetime, directory, prefix and suffix
=======
\item{suffix}{suffix dat bijvoorbeeld een db id kan bevatten}
}
\value{
file connectie
}
\description{
Creert logbestand (opent connectie)
>>>>>>> d798110a5b4d52f71d723703657283586349b6aa
}
