% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfile_functions.R
\name{logfile_write}
\alias{logfile_write}
\title{Write to logfile (sink)}
\usage{
logfile_write(logfile, new = FALSE, close = FALSE)
}
\arguments{
\item{logfile}{file to sink the output to}

\item{new}{set to TRUE if a new file is to be created}

\item{close}{close the output diversion?}
}
\value{
NULL, output redirected
}
\description{
Write to logfile (sink)
}
