% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elc_qc_specs_functions.R
\name{make_spec_queries}
\alias{make_spec_queries}
\title{Maak insert, update, delete en select queries voor de C_QC_LIMITS tabel}
\usage{
make_spec_queries(
  label,
  active = "F",
  qc_sample,
  qc_version,
  analysis,
  component,
  product,
  product_version,
  sampling_point,
  theo_value = NULL,
  theo_sd = NULL,
  meas_value = NULL,
  meas_sd = NULL
)
}
\arguments{
\item{label}{label om de qc limieten op de selecteren}

\item{active}{"T" indien in gebruik, "F" dan kunnen die niet meer gebruikt worden}

\item{qc_sample}{naam van qc_sample in qc_samples tabel (= PRODUCT_GRADE)}

\item{qc_version}{versie van qc_sample}

\item{analysis}{analyse waarvoor de limieten gelden}

\item{component}{component waarvoor limieten gelden}

\item{product}{product waar QC_SAMPLE bij hort}

\item{product_version}{productversie die bij de limieten hoort}

\item{sampling_point}{sampling_point zoals in SAMPLING_POINT em PRODUCT_SPECS tabel}

\item{theo_value}{theoretische waarde}

\item{theo_sd}{theoretische sd (zal vaak 0 zijn)}

\item{meas_value}{gemeten gemiddelde waarde}

\item{meas_sd}{gemeten standaarddeviatie, waarop de 3s grenzen worden berekend}
}
\value{
lijst met 4 elementen (select, update, delete, insert) die elk de respectievelijke query bevatten
}
\description{
Maak insert, update, delete en select queries voor de C_QC_LIMITS tabel
}
