% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_importfile_parse.R
\name{move_batch_importfile}
\alias{move_batch_importfile}
\title{Verplaats file naar lims importlocatie}
\usage{
move_batch_importfile(
  data,
  source_path,
  source_file,
  batch_info,
  scheduler_base_dir
)
}
\arguments{
\item{data}{dataset die als textbestand naar de juiste locatie wordt verplaatst}

\item{source_path}{locatie van het bronbestand}

\item{source_file}{bronbestand inclusief padnaam}

\item{batch_info}{batch_info dataset met de batchinformatie die minstens de kolommen template, name, c_import_routine, c_import_sheet, c_root_dir en group_name bevat}

\item{scheduler_base_dir}{directory waaronder de structuur staat waar de file naartoe geschreven moet worden}
}
\value{
geen return, wel verplaatste bestanden
}
\description{
Verplaats file naar lims importlocatie
}
