% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_create_report.R
\name{plate_create_report}
\alias{plate_create_report}
\title{Maak een plaatlayout rapport aan}
\usage{
plate_create_report(data, Capilar = LETTERS[1:8], Lane = 1:12)
}
\arguments{
\item{data}{datasset komende uit \link[inbolimsintern]{plate_setup_source}}

\item{Capilar}{de codes van de plaatkolommen, typisch A tot F}

\item{Lane}{de nummers van de plaatrijen, typisch 1 tot 12}
}
\value{
dataset klaar om te exporteren naar excel of naar de C_DNA_RUN_REPORT_RESULTS tabel
}
\description{
Maak een plaatlayout rapport aan
}
