% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_read_dna_run.R
\name{plate_read_dna_run}
\alias{plate_read_dna_run}
\title{Lees data van een DNA run in om daar later een plaat layout van te maken}
\usage{
plate_read_dna_run(conn, run)
}
\arguments{
\item{conn}{de connectie met de databank}

\item{run}{de DNA_RUN_ID vanuit het lims systeem}
}
\value{
dataset containig PLATE, ROW_NUMBER, COLUMN_NUMBER, SAMPLE_NUMBER, PARENT_SAMPLE, DNA, MilliQ
}
\description{
Lees data van een DNA run in om daar later een plaat layout van te maken
}
