% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_setup_source.R
\name{plate_setup_source}
\alias{plate_setup_source}
\title{Brondata voor plaatopvulschema klaarzetten}
\usage{
plate_setup_source(data)
}
\arguments{
\item{data}{dataset die uit de tabel C_DNA_RUN_REPORT komt}
}
\value{
design dataset voor de plaatlayout
}
\description{
Brondata voor plaatopvulschema klaarzetten
}
