% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber_batch_file_grabber.R
\name{plumber_batch_file_grabber}
\alias{plumber_batch_file_grabber}
\title{Batch file grabber}
\usage{
plumber_batch_file_grabber(
  env = "PRD",
  call_id = 0,
  user_name = "",
  scheduler_path =
    "\\\\\\\\Inbo-limsbg-prd-labware8.inbo.be\\\\LABO_FS_PRD\\\\ANA\\\\_SCHEDULER"
)
}
\arguments{
\item{env}{PRD, UAT of DEV}

\item{scheduler_path}{path waar scheduler de files zoekt}
}
\value{
aantal geconverteerde files
}
\description{
Batch file grabber
}
