% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber_qc_chart_simple.R
\name{plumber_qc_chart_simple}
\alias{plumber_qc_chart_simple}
\title{Generate a quick and simple html page with the qc plots}
\usage{
plumber_qc_chart_simple(env, call_id, user_name, sqlfile, htmlfile, maxpoints)
}
\arguments{
\item{env}{environment PRD, UAT or DEV}

\item{call_id}{(int) ID for the call}

\item{user_name}{lims user name}

\item{sqlfile}{where the sql code is housed}

\item{htmlfile}{the resulting html file}

\item{maxpoints}{the maximum amount of points to go back to per chart}
}
\value{
location of the html file
}
\description{
Generate a quick and simple html page with the qc plots
}
