% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_importfile_parse.R
\name{process_file_generic}
\alias{process_file_generic}
\title{Kopieer de data naar een tsv en verplaats de originele file}
\usage{
process_file_generic(
  data,
  source_file,
  target_location = ".",
  move_location = "_FINISHED"
)
}
\arguments{
\item{data}{dataset die ingelezen is}

\item{source_file}{het volledig path met bestandsnaam (nodig om te verplaatsen en de file naamgeving)}

\item{target_location}{het volledig path naar de doeldirectory, ofwel "." voor op dezelfde plaats te blijven}

\item{move_location}{het volledig  path naar de verplaatsdirectory, ofwel "_FINISHED" om relatief van de file in de _FINISHED directory te schrijven}
}
\value{
succes status
}
\description{
Kopieer de data naar een tsv en verplaats de originele file
}
