% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filegrabber_process_proj_reg.R
\name{process_proj_reg}
\alias{process_proj_reg}
\title{Process files ontrent projectregistratie}
\usage{
process_proj_reg(source_path, source_file, target_path, finish_path, sheet)
}
\arguments{
\item{source_path}{pad waar het bronbestand te vinden is}

\item{source_file}{naam van het bronbestand (xlsx)}

\item{target_path}{het pad waar de tsv naartoe geschreven moet worden}

\item{finish_path}{het pad waar de excel file heen noet na deze processing}

\item{sheet}{naam van het tabblad dat ingelezen moet worden}
}
\value{
gecreerde en verplaatse bestanden
}
\description{
Process files ontrent projectregistratie
}
