% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filegrabber_process_samp_reg.R
\name{process_samp_reg}
\alias{process_samp_reg}
\title{Process files ontrent staalregistratie}
\usage{
process_samp_reg(
  source_path,
  source_file,
  target_path,
  finish_path,
  sheet_samp,
  sheet_ana
)
}
\arguments{
\item{source_path}{pad waar het bronbestand te vinden is}

\item{source_file}{naam van het bronbestand (xlsx)}

\item{target_path}{het pad waar de tsv naartoe geschreven moet worden}

\item{finish_path}{het pad waar de excel file heen noet na deze processing}

\item{sheet_samp}{naam van het tabblad met de staalinformatie}

\item{sheet_ana}{naam van het tabblad met de te loggen analyses}
}
\value{
gecreerde en verplaatse bestanden
}
\description{
Process files ontrent staalregistratie
}
