% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elc_qcc_lims_shewhart_rules.R
\name{qcc_rule01}
\alias{qcc_rule01}
\title{Rule1: point outside 3sigma}
\usage{
qcc_rule01(x, lcl_3s, ucl_3s, run = 1)
}
\arguments{
\item{x}{numeric vactor of lab results}

\item{lcl_3s}{lower limit for violation}

\item{ucl_3s}{upper limit for violation}

\item{run}{should always 1}
}
\value{
TRUE if rule violated
}
\description{
Rule1: point outside 3sigma
}
