% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcc_lims_shewhart_rules.R
\name{qcc_rule02}
\alias{qcc_rule02}
\title{Rule2: 9 consecutive points on the same side of the center}
\usage{
qcc_rule02(x, center, run = 9)
}
\arguments{
\item{x}{numeric vector of lab results}

\item{center}{center line off which the side is calculated}

\item{run}{how many times the same side before rule violation is triggered}
}
\value{
TRUE if rule violation
}
\description{
Rule2: 9 consecutive points on the same side of the center
}
