% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcc_lims_shewhart_rules.R
\name{qcc_rule03}
\alias{qcc_rule03}
\title{Rule3: 6 consecutive points steadily increasing or decreasing}
\usage{
qcc_rule03(x, run = 6)
}
\arguments{
\item{x}{numeric vector of lab results}

\item{run}{how long monotonous increase or decrease before rule violation}
}
\value{
TRUE if rule violation
}
\description{
Rule3: 6 consecutive points steadily increasing or decreasing
}
