% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcc_lims_shewhart_rules.R
\name{qcc_rule04}
\alias{qcc_rule04}
\title{Rule4: 14 consecutive points alternating up and down}
\usage{
qcc_rule04(x, run = 14)
}
\arguments{
\item{x}{numeric vector of lab results}

\item{run}{number of alternating up and down before rule violation}
}
\value{
TRUE if rule violation
}
\description{
Rule4: 14 consecutive points alternating up and down
}
