% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elc_qcc_lims_shewhart_rules.R
\name{qcc_rule05b}
\alias{qcc_rule05b}
\title{Rule5b: variation on Rule5: 2 consecutive points more than 2 sigma on same side}
\usage{
qcc_rule05b(x, lcl_2s, ucl_2s, run = 2)
}
\arguments{
\item{x}{numeric vector of lab results}

\item{lcl_2s}{2-sigma lower limit}

\item{ucl_2s}{2-sigma upper limit}

\item{run}{(run - 1) violations on run more than 2 sigma in the same direction}
}
\value{
TRUE if rule violated
}
\description{
Rule5b: variation on Rule5: 2 consecutive points more than 2 sigma on same side
}
