% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcc_lims_shewhart_rules.R
\name{qcc_rule07}
\alias{qcc_rule07}
\title{Rule7: 15 consecutive points are within 1 sigma of the center line}
\usage{
qcc_rule07(x, lcl_1s, ucl_1s, run = 15)
}
\arguments{
\item{x}{numeric vector of lab results}

\item{lcl_1s}{lower 1-sigma limit}

\item{ucl_1s}{upper 1-sigma limit}

\item{run}{how many that meet the criterion before violation is triggered}
}
\value{
TRUE if rule violation
}
\description{
Rule7: 15 consecutive points are within 1 sigma of the center line
}
