% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcc_lims_shewhart_rules.R
\name{qcc_rule08}
\alias{qcc_rule08}
\title{Rule8: 8 consecutive points on either side of the center line not within 1 sigma}
\usage{
qcc_rule08(x, lcl_1s, ucl_1s, run = 8)
}
\arguments{
\item{x}{numeric vector of lab results}

\item{lcl_1s}{1-sigma lower limit}

\item{ucl_1s}{1-sigma upper limit}

\item{run}{how many before violation is triggerd}
}
\value{
TRUE if violation
}
\description{
Rule8: 8 consecutive points on either side of the center line not within 1 sigma
}
