% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzx_r_session_setup.R
\name{r_session_setup}
\alias{r_session_setup}
\title{Prepare a remote script session}
\usage{
r_session_setup(
  args = commandArgs(trailingOnly = TRUE),
  test_mode = FALSE,
  cred_file = "dbcredentials.txt",
  logfile = "lims_rsession.log"
)
}
\arguments{
\item{args}{defaults to commandArgs(trailingOnly = TRUE), but user can specify a custom character vector of arguments. The first 3 arguments must be dbodbc, call_id, and user}

\item{test_mode}{if TRUE the credentials are read from the cred_file}

\item{cred_file}{Path to the credentials file containing
database connection information. Used only in interactive/test mode.
The file should contain at least three lines containing
and just one column containing: data source, username, and password
in this exact order
Default is "dbcredentials.txt".}

\item{logfile}{logfile to write to, because this routine is ran before the databae connection is established and the logs can be saved there}
}
\value{
A list containing essential objects for the script:
  \itemize{
    \item \code{conn}: The DBI database connection object.
    \item \code{params}: A list of parameters read from the database.
  }
}
\description{
This high-level function handles all boilerplate setup for scripts run via Rscript.
It prepares the session, connects to the LIMS database, reads parameters,
and sets up logging and pandoc paths.
}
