% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read_db_arguments.R
\name{read_db_arguments}
\alias{read_db_arguments}
\title{Lees argumenten uit de database}
\usage{
read_db_arguments(conn, call_id)
}
\arguments{
\item{conn}{connection object to the LIMS database}

\item{call_id}{numeric identifier of the function you want to get the arguments from}
}
\value{
named list with arguments
}
\description{
Lees argumenten uit de database
}
\examples{
\dontrun{
conn = lims_db_connect(uid = "ikke", pwd = "123456")
read_db_arguments(conn, call_id = 10)
}
}
