% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read_db_log.R
\name{read_db_log}
\alias{read_db_log}
\title{Haal de status op van het lopen van het R script}
\usage{
read_db_log(conn, call_id)
}
\arguments{
\item{conn}{db connection object}

\item{call_id}{integer specifying which call is run}
}
\value{
dataset with current status of the table
}
\description{
Haal de status op van het lopen van het R script
}
