% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Verwijder waarden buiten een vooropgesteld aantal sigmas}
\usage{
remove_outliers(x, nsigma = 3)
}
\arguments{
\item{x}{vector of numerical values}

\item{nsigma}{numeric value containing the n-sigma values should be removed}
}
\value{
vector without outliers
}
\description{
Verwijder waarden buiten een vooropgesteld aantal sigmas
}
\examples{
x <- rnorm(100000)
length(remove_outliers(x))
}
