% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_chart_functions.R
\name{run_chart_breaks}
\alias{run_chart_breaks}
\title{Labels voor x as run chart}
\usage{
run_chart_breaks(numeric_labels, label_values, n_max = 40)
}
\arguments{
\item{numeric_labels}{numerieke labels (zoals ze gezien worden door de x as)}

\item{label_values}{vector met labels die horen bij de numeric_labels}

\item{n_max}{maximum aantal breaks}
}
\value{
lijst met 2 componten met gelijke lengte: value en label
}
\description{
Labels voor x as run chart
}
