% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_chart_functions.R
\name{run_chart_query}
\alias{run_chart_query}
\title{Basisquery voor run chart}
\usage{
run_chart_query(params)
}
\arguments{
\item{params}{data.frame met kolommen ARG_NAME (PROJECT, Y1 + optioneel Y2,Y3,Y4,Y5) en VALUE}
}
\value{
query string
}
\description{
Basisquery voor run chart
}
