% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcc_select_control_samples.R
\name{select_control_samples}
\alias{select_control_samples}
\title{Selecteer de waarden die als basis voor de controlekaart dienen}
\usage{
select_control_samples(conn, num, batch, analysis, components,
  start = "1900-01-01", end = "2200-01-01")
}
\arguments{
\item{conn}{db connection}

\item{num}{number of samples to keep or all samples of a year if coded as JD2017 or all samples from the last D days, coded starting with a D, so num can be like: 30 or JD2017 or D60}

\item{batch}{the name of the batch, which will be used to check which control samples are returned}

\item{analysis}{the name of the analysis}

\item{components}{a vector of one or more components within the analysis}

\item{start}{the earliest startdate coded as "YYYY-mm-dd". Only if num is a numeric value}

\item{end}{the latest enddate coded as "YYYY-mm-dd". Only used if num is a numeric value}
}
\value{
a dataset with all lab information (SAMPLE_NAME, SAMPLE_NUMBER, TEXT_ID, ANALYSIS, NAME, BATCH, ENTRY, C_DATE_BATCHRUN, ENTERED_ON, Nr, CV)
}
\description{
Selecteer de waarden die als basis voor de controlekaart dienen
}
\examples{
\dontrun{
batch = "IC_AN-190430-1"
analysis = "IC_ANIONEN"
components = c("NO2", "NO3")
testdf <- select_control_samples(conn, num = 30, batch, analysis, components)
}
}
