% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_set_edna_plate_positions.R
\name{set_edna_plate_positions}
\alias{set_edna_plate_positions}
\title{Data verder klaarzettten voor export naar lims platen}
\usage{
set_edna_plate_positions(
  data,
  n_plates,
  plate_size = 96,
  column_size = 8,
  qcm_pos = 38,
  n_blank = 3
)
}
\arguments{
\item{data}{dataset komende uit de functie get_edna_plate_data}

\item{n_plates}{aantal platen die gevuld moeten worden}

\item{plate_size}{grootte van de plaat (standaard 96)}

\item{column_size}{aantal lanes (standaard 8 voor een 96 wells plate)}

\item{qcm_pos}{posities van de qc_methods}

\item{n_blank}{aantal blanco's per plaat}
}
\value{
dataset die kan weggeschreven worden in C_DNA_EXTRACTION waar Lims dan verder mee aan de slag kan
}
\description{
Data verder klaarzettten voor export naar lims platen
}
