% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzy_setup_lims_environment.R
\name{setup_lims_environment}
\alias{setup_lims_environment}
\title{Run the LIMS script setup process.}
\usage{
setup_lims_environment()
}
\description{
This function iterates through a predefined list of LIMS script paths.
It stops and updates the first accessible path it finds.
}
\keyword{internal}
