% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elc_qc_specs_functions.R
\name{update_qc_limits}
\alias{update_qc_limits}
\title{Update de C_QC_LIMITS tabel}
\usage{
update_qc_limits(conn, data)
}
\arguments{
\item{conn}{connectie naar db}

\item{data}{dataset die alle records bevat die geupdated moeten worden}
}
\value{
invisible
}
\description{
Update de C_QC_LIMITS tabel
}
