% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_write_archive_data.R
\name{write_archive_data}
\alias{write_archive_data}
\title{Schrijf archiefkaartdata zeg naar de database}
\usage{
write_archive_data(conn, data)
}
\arguments{
\item{conn}{DBI connectieobject naar database}

\item{data}{de weg te schrijven data}
}
\value{
written records in database table C_CTR_ARCHIVE
}
\description{
Schrijf archiefkaartdata zeg naar de database
}
