% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write_db_log.R
\name{write_db_log}
\alias{write_db_log}
\title{Write a Log Entry to the Database}
\usage{
write_db_log(
  conn,
  call_id,
  status,
  message = "",
  extra = "",
  user = "",
  print = T
)
}
\arguments{
\item{conn}{A DBI database connection object.}

\item{call_id}{A unique identifier for the specific script execution.}

\item{status}{A character code representing the status. It's recommended to use
a consistent set of codes, for example:
\itemize{
  \item \code{START}: The process has started.
  \item \code{SUCCESS}: The process completed without errors.
  \item \code{FAIL}: The process terminated with an error.
  \item \code{INFO}: A generic informational message.
}}

\item{message}{An optional character string providing additional details or an
error message. Defaults to an empty string.}

\item{extra}{Extra information that is printed (solely on the console) after the message}

\item{user}{optional the user name from whom this log is created}

\item{print}{If TRUE also a message is printed on the console}
}
\value{
Invisibly returns the number of rows affected by the insert statement
  (typically 1 on success). The return value is invisible to prevent it
  from printing to the console.
}
\description{
Inserts a new record into the `C_RSCRIPT_LOG` table to log the status of a
script execution.
}
\examples{
\dontrun{
# Assuming 'conn' is an active database connection
call_id <- 12345

# Log the start of a script
write_db_log(conn, call_id, "START", "Processing started for monthly report.")

# Log a successful completion
write_db_log(conn, call_id, "SUCCESS")

# Log a failure
write_db_log(conn, call_id, "FAIL", "Error in data aggregation step.")
}
}
