% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elc_functions_obsolete.R
\name{compare_with_tf}
\alias{compare_with_tf}
\title{Voer een t en F test uit (OUDE MANIER)}
\usage{
compare_with_tf(x, years)
}
\arguments{
\item{x}{waarden}

\item{years}{verschillende periodes}
}
\value{
dataset met t en F test resultaten
}
\description{
Voer een t en F test uit (OUDE MANIER)
}
