% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{datetime_text}
\alias{datetime_text}
\title{Get current datetime as character variable}
\usage{
datetime_text()
}
\value{
character string yyyymmdd_hhmmss
}
\description{
Get current datetime as character variable
}
