% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elc_htmldata.R
\name{elc_htmldata}
\alias{elc_htmldata}
\title{Maak data aan voor elc shewhart html}
\usage{
elc_htmldata(
  plotdata,
  colors = c("lightblue3", "green4", "orange", "red", "blue4"),
  base_size = 1.5,
  expected_value = NULL,
  expected_sd = NULL,
  digits = 5,
  check_rules = TRUE,
  reverse_table_order = TRUE
)
}
\arguments{
\item{plotdata}{dataset typisch van get_ELC_data}

\item{colors}{vector van 5 kleuren die gebruikt worden in plot}

\item{base_size}{basis puntgrootte in de plot}

\item{expected_value}{indien NULL berekend uit de data, anders wordt deze gebruikt}

\item{expected_sd}{indien NULL berekend uit de data, anders wordt deze gebruikt}

\item{digits}{aantal digits te printen in de html}

\item{check_rules}{moeten de grenzen berekend worden of zitten die al als kolom in de data}

\item{reverse_table_order}{Moet de data van recent naar oud geschikt worden in de html tabel}
}
\value{
lijst met 4 datasets: plot, borders, summary en tabel
}
\description{
Maak data aan voor elc shewhart html
}
