% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elc_functions.R
\name{ELC_shewhart_plot}
\alias{ELC_shewhart_plot}
\title{Plot ELC shewhart plot}
\usage{
ELC_shewhart_plot(subdata, borders, base_color = "lightblue3")
}
\arguments{
\item{subdata}{data om te plotten met kolommen batchnr, waarde, color, size}

\item{borders}{dataframe met de kolommen val en color om horizontale lijnen voor de s-grenzen te trekken}

\item{base_color}{basiskleur in de grafieken}
}
\value{
ggplot2 object
}
\description{
Plot ELC shewhart plot
}
