% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_importfile_parse.R
\name{get_data_from_importfile}
\alias{get_data_from_importfile}
\title{Krijg de informatie uit het excel tabblad of txt of csv bestand}
\usage{
get_data_from_importfile(path, batch_info)
}
\arguments{
\item{path}{naam van de bestandsnaam inclusief het pad ernaar}

\item{batch_info}{dataset met de batchinformatie die minstens de kolommen template, name, c_import_routine, c_import_sheet, c_root_dir, en group_name, bevat}
}
\value{
dataset et alle importgegevens
}
\description{
Krijg de informatie uit het excel tabblad of txt of csv bestand
}
