% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_limsdb_connect.R
\name{limsdb_connect}
\alias{limsdb_connect}
\title{Create a connection to the LIMS database}
\usage{
limsdb_connect(env = "PRD", connectlist = NULL)
}
\arguments{
\item{env}{which environment PRD, UAT, DEV or LW7}

\item{connectlist}{list containinf connection information, should at least have the elements host, dsn, uid and pwd
It requires a correctly configured .Renviron file to be present in the
user's home directory or the project's working directory.}
}
\value{
A DBI database connection object.
}
\description{
This function automatically detects the environment (DEV, UAT, PRD, LW7),
retrieves the appropriate credentials using Sys.getenv(), and establishes
a database connection.
}
