% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_limsdb_connect.R
\name{limsdb_connect}
\alias{limsdb_connect}
\title{Title}
\usage{
limsdb_connect(server = "inbo-sql07-prd.inbo.be",
  database = "D0015_00_Lims", uid, pwd, connectlist = NULL)
}
\arguments{
\item{server}{name of the LIMS server}

\item{database}{name of the LIMS database}

\item{uid}{application username (that as writing rights on the db, so this is not you username)}

\item{pwd}{applciation username password}

\item{connectlist}{list containing dsn (server name), uid (user id) and pwd (password). Optionally it can also contain db. It overrides the values of the other arguments}
}
\value{
db connection
}
\description{
Title
}
\examples{
\dontrun{
lims_db_connect(uid = "me", pwd = "123456") #should not work
}
}
