% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_limsdwh_connect.R
\name{limsdwh_connect}
\alias{limsdwh_connect}
\title{Connect to lims datawarehouse}
\usage{
limsdwh_connect(deployment = "prd", use_rodbc = FALSE, uidpwd = "")
}
\arguments{
\item{deployment}{'uat' for test server 'prd' for production server}

\item{use_rodbc}{use the rodbc library (or the DBI when FALSE)}

\item{uidpwd}{not used for the moment}
}
\description{
Connect to the lims datawarehouse using a trusted connection as INBO staff member
}
