% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfile_functions.R
\name{logfile_start}
\alias{logfile_start}
\title{Make name of log file based on current datetime, directory, prefix and suffix}
\usage{
logfile_start(path = "<auto>", prefix = "", suffix = "")
}
\arguments{
\item{path}{directory to write log file}

\item{prefix}{prefix to the file to make it recognisable}

\item{suffix}{suffix tot the file to make it recognisable}
}
\value{
file connection
}
\description{
Make name of log file based on current datetime, directory, prefix and suffix
}
