% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfile_functions.R
\name{logfile_start}
\alias{logfile_start}
\title{Haal de Rscript logdirectory uit de LIMS databank}
\usage{
logfile_start(path = "<auto>", prefix = "", suffix = "")
}
\arguments{
\item{path}{directory to write log file}

\item{prefix}{prefix to the file to make it recognisable}

\item{suffix}{suffix tot the file to make it recognisable}

\item{conn}{connectie naar de databank}
}
\value{
logdirectory (indien niet gevonden NULL)

file connection
}
\description{
Haal de Rscript logdirectory uit de LIMS databank
}
