% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfile_functions.R
\name{logfile_start}
\alias{logfile_start}
\title{Creert logbestand (opent connectie)}
\usage{
logfile_start(path = "<auto>", prefix = "", suffix = "")
}
\arguments{
\item{path}{directory waar de logfile wordt aangemaakt . <auto> gebruikt de default}

\item{prefix}{vorvoegsel in de file, bijvoorbeeld welk script gebruikt werd}

\item{suffix}{suffix dat bijvoorbeeld een db id kan bevatten}
}
\value{
file connectie
}
\description{
Creert logbestand (opent connectie)
}
