% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_print_labels.R
\name{print_lims_labels}
\alias{print_lims_labels}
\title{Master function to handle API previews or direct network printing.}
\usage{
print_lims_labels(
  dataset,
  printer_config,
  format_lines,
  mode = c("api", "real"),
  abort_real_show_payload = FALSE,
  format = c("png", "pdf"),
  output_path = "C:/Labels/label_output.pdf",
  batch_mode = FALSE,
  batch_size = 8
)
}
\arguments{
\item{dataset}{Dataframe. The sample data to print.}

\item{printer_config}{List. Hardware settings.}

\item{format_lines}{Dataframe. Label design definitions.}

\item{mode}{Character. "api" for preview, "real" for physical print.}

\item{abort_real_show_payload}{stop just before printing on the printer and show the payload that would be sent}

\item{format}{Character. "png" (single) or "pdf" (batch) for API previews.}

\item{output_path}{Character. Path to save the output file.}

\item{batch_mode}{for in the future if working with batch jobs instead of label per label}

\item{batch_sie}{how many labels per batch}
}
\description{
Master function to handle API previews or direct network printing.
}
\examples{
\dontrun{
conn <- limsdb_connect()
p_cfg <- get_printer_config(conn, "G35_LABO_ZEBRA")
f_lns <- get_label_format(conn, 'PROJ_LABEL_EXTID')
f_lns <- adjust_label_positions(f_lns, target_height_mm = p_cfg$LENGTH)
samples <- data.frame(SAMPLE_ID = "extern id van dit", TEXT_ID="D26-123456-01", PROJECT="V-25W100-05")

# Preview as PDF
print_lims_labels(samples, p_cfg, f_lns, mode = "api", format = "pdf")
}
}
