% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filegrabber_process_field_form.R
\name{process_field_form}
\alias{process_field_form}
\title{Process files omtrent inlezen veldformulier}
\usage{
process_field_form(
  source_path,
  source_file,
  target_path,
  finish_path,
  sheet,
  digits = 12
)
}
\arguments{
\item{source_path}{pad waar het bronbestand te vinden is}

\item{source_file}{naam van het bronbestand (xlsx)}

\item{target_path}{het pad waar de tsv naartoe geschreven moet worden}

\item{finish_path}{het pad waar de excel file heen noet na deze processing}

\item{sheet}{naam van het tabblad dat ingelezen moet worden}

\item{digits}{hoeveel digits behouden bij het bewaren van nummerieke waarden}
}
\value{
gecreerde en verplaatse bestanden
}
\description{
Process files omtrent inlezen veldformulier
}
