% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzx_r_session_setup.R
\name{r_session_setup}
\alias{r_session_setup}
\title{Prepare a remote script session}
\usage{
r_session_setup(
  call_id,
  odbc,
  args = commandArgs(trailingOnly = TRUE),
  cred_file = "dbcredentials.txt",
  logfile = "lims_rsession.log"
)
}
\arguments{
\item{call_id}{The unique identifier for the script call, passed via command line as commandArgs()[2]}

\item{odbc}{The odbc that is given as commandArgs()[1] to define environment}

\item{args}{defaults to commandArgs(trailingOnly = TRUE), but user can specify a custom character vector of arguments. The first 4 arguments must be dbodbc, dbuid, dbpwd and call_id, optionally the fifth should be user}

\item{cred_file}{Path to the credentials file containing
database connection information. Used only in interactive/test mode.
The file should contain at least three lines containing
and just one column containing: data source, username, and password
in this exact order
Default is "dbcredentials.txt".}
}
\value{
A list containing essential objects for the script:
  \itemize{
    \item \code{conn}: The DBI database connection object.
    \item \code{params}: A list of parameters read from the database.
  }
}
\description{
This high-level function handles all boilerplate setup for scripts run via Rscript.
It prepares the session, connects to the LIMS database, reads parameters,
and sets up logging and pandoc paths.
}
