% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read_db_credentials.R
\name{read_db_credentials}
\alias{read_db_credentials}
\title{Lees de data credentials van de LIMS hoofddatabank}
\usage{
read_db_credentials(file = "dbcredentials.txt")
}
\arguments{
\item{file}{bestandsnaam waaruit de credentials geschreven worden}
}
\value{
list with 3 elements, the data source, user-id and paswoord of the LIMS Main DB
}
\description{
Lees de credentials in voor de database vanuit een bestand. De eerste regel bevat de data source, de tweede regel de username, en de derde regel het passwoord
}
\examples{
read_db_credentials(system.file("extdata", "dbcredentials.txt", package = "inbolimsintern"))
}
