% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_statistics.R
\name{results_statistics}
\alias{results_statistics}
\title{Make a csv and histograms about data statistics}
\usage{
results_statistics(path, basename)
}
\arguments{
\item{path}{directory to store file in}

\item{basename}{base file name to store csv and png}
}
\value{
a csv and png file
}
\description{
Make a csv and histograms about data statistics
}
