% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write_db_log.R
\name{write_db_log}
\alias{write_db_log}
\title{Write a log message to the database.}
\usage{
write_db_log(
  message,
  status,
  call_id = NULL,
  user_name = NULL,
  conn = NULL,
  extra = "",
  print_log = TRUE
)
}
\arguments{
\item{message}{The log message string to be written. (Required)}

\item{status}{The status of the log entry (e.g., "I", "P" (in progress), "E" (error), "C" (completed)). (Required)}

\item{call_id}{A unique identifier for the script execution. If NULL,
the function will look for a 'call_id' variable in the parent frame.}

\item{user_name}{The username associated with the log entry. If NULL,
the function will look for a 'user_name' variable in the parent frame.}

\item{conn}{The database connection object. If NULL, the function will
look for a 'conn' variable in the parent frame.}

\item{extra}{An optional string to append to the message when printing to the
console. Defaults to an empty string.}

\item{print_log}{A logical value indicating whether to print the message to the
console. Defaults to TRUE.}
}
\value{
Invisibly returns the number of rows affected by the SQL execution and writes a message to the console
}
\description{
This function inserts a log entry into the C_RSCRIPT_LOG table.
The 'message' and 'status' are required. Other parameters like 'call_id',
'user', and 'conn' are optional; if not provided, the function will attempt
to retrieve them from the parent environment.
}
\examples{
\dontrun{
# Assuming 'my_conn', 'current_user', and 'run_id' exist in the environment
conn <- my_conn
user_name <- current_user
call_id <- run_id

# Now you can call the function with only the required arguments
write_db_log("Starting the process", "P")
# ... some processing ...
write_db_log("Process finished successfully", "C")
}
}
